cycles, add test game.

+++ src/com/ferdin/nescpu/DemoNESTest.java
@@ -6,7 +6,7 @@ public class DemoNESTest {
     @Test
     void test_0xA9_lda_immediate_load_data(){
         DemoNES nes = new DemoNES();
-        nes.loadAndRun(new byte[] {(byte) 0xA9, (byte) 0x05, (byte) 0x00});
+        nes.loadAndRun(new int[] { 0xA9, 0x05, 0x00 });
 
         assertEquals(0x05, nes.registerA);
         assertEquals(0, nes.status & 0b0000_0010); // Zero flag should be clear
@@ -16,8 +16,8 @@ public class DemoNESTest {
     void test_lda_from_memory(){
         DemoNES nes = new DemoNES();
         nes.memWrite(0x10,(byte) 0x55);
-        nes.loadAndRun(new byte[] {
-            (byte) 0xa5, (byte) 0x10, (byte) 0x00
+        nes.loadAndRun(new int[] {
+            0xa5, 0x10, 0x00
         });
         assertEquals(nes.registerA, 0x55);
     }
@@ -25,10 +25,10 @@ public class DemoNESTest {
     void test_0xA9_lda_zero_flag() {
         DemoNES nes = new DemoNES();
 
-        nes.loadAndRun(new byte[] {
-                (byte) 0xA9,
-                (byte) 0x00,
-                (byte) 0x00
+        nes.loadAndRun(new int[] {
+                0xA9,
+                0x00,
+                0x00
         });
 
         assertEquals(0b10, nes.status & 0b0000_0010); // Zero flag should be set
@@ -36,20 +36,20 @@ public class DemoNESTest {
     @Test
     void test_0xaa_tax_move_a_to_x(){
         DemoNES nes = new DemoNES();
-        nes.loadAndRun(new byte[] {
-            (byte) 0xA9, (byte) 0x0A,  // LDA #$0A (load 10 into A)
-            (byte) 0xAA,                // TAX (transfer A to X)
-            (byte) 0x00                 // BRK
+        nes.loadAndRun(new int[] {
+            0xA9, 0x0A,  // LDA #$0A (load 10 into A)
+            0xAA,        // TAX (transfer A to X)
+            0x00         // BRK
         });
         assertEquals(10, nes.registerX);
     }
     @Test
     void test_5_ops_working_together(){
         DemoNES nes = new DemoNES();
-        nes.loadAndRun(new byte[]{
-                (byte) 0xA9, (byte) 0xC0, 
-                (byte) 0xAA,              
-                (byte) 0xE8, (byte) 0x00           
+        nes.loadAndRun(new int[]{
+                0xA9, 0xC0, 
+                0xAA,              
+                0xE8, 0x00           
         });
 
         assertEquals(0xC1, nes.registerX & 0xFF);
@@ -58,43 +58,43 @@ public class DemoNESTest {
     void test_inx_overflow(){
         DemoNES nes = new DemoNES();
         nes.registerX = 0xFF;
-        nes.loadAndRun(new byte[]{
-            (byte) 0xE8, (byte) 0xE8, (byte) 0x00
+        nes.loadAndRun(new int[]{
+            0xE8, 0xE8, 0x00
         });
         assertEquals(2, nes.registerX & 0xFF);
     }
     @Test
     void test_bcc_branch_taken(){
         DemoNES nes = new DemoNES();
-        nes.loadAndRun(new byte[]{
-            (byte) 0x90, (byte) 0x02, // BCC +2
-            (byte) 0xA9, (byte) 0x01, // LDA #$01 (should be skipped)
-            (byte) 0xA9, (byte) 0x02, // LDA #$02 (should be executed)
-            (byte) 0x00
+        nes.loadAndRun(new int[]{
+            0x90, 0x02, // BCC +2
+            0xA9, 0x01, // LDA #$01 (should be skipped)
+            0xA9, 0x02, // LDA #$02 (should be executed)
+            0x00
         });
         assertEquals(0x02, nes.registerA);
     }
     @Test
     void test_bcs_branch_taken(){
         DemoNES nes = new DemoNES();
-        nes.loadAndRun(new byte[]{
-            (byte) 0x38, // SEC - Set Carry Flag
-            (byte) 0xB0, (byte) 0x02, // BCC +2
-            (byte) 0xA9, (byte) 0x01, // LDA #$01 (should be skipped)
-            (byte) 0xA9, (byte) 0x02, // LDA #$02 (should be executed)
-            (byte) 0x00
+        nes.loadAndRun(new int[]{
+            0x38, // SEC - Set Carry Flag
+            0xB0, 0x02, // BCS +2
+            0xA9, 0x01, // LDA #$01 (should be skipped)
+            0xA9, 0x02, // LDA #$02 (should be executed)
+            0x00
         });
         assertEquals(0x02, nes.registerA);
     }
     @Test
     void test_bcs_branch_not_taken(){
         DemoNES nes = new DemoNES();
-        nes.loadAndRun(new byte[]{
-            (byte) 0xB0, (byte) 0x02, // BCS +2 (not taken)
-            (byte) 0xA9, (byte) 0x01, // LDA #$01 (should be executed)
-            (byte) 0x00,              // BRK (stop here)
-            (byte) 0xA9, (byte) 0x02, // LDA #$02 (should be skipped)
-            (byte) 0x00
+        nes.loadAndRun(new int[]{
+            0xB0, 0x02, // BCS +2 (not taken)
+            0xA9, 0x01, // LDA #$01 (should be executed)
+            0x00,              // BRK (stop here)
+            0xA9, 0x02, // LDA #$02 (should be skipped)
+            0x00
         });
         assertEquals(0x01, nes.registerA);
     }
@@ -102,11 +102,11 @@ public class DemoNESTest {
     void test_bit(){
         DemoNES nes = new DemoNES();
         nes.registerA = 0b00000001;
-        nes.memWrite(0x10, (byte)0b11000000);
-        nes.loadAndRun(new byte[]{
-            (byte)0x24,
-            (byte)0x10,
-            (byte)0x00
+        nes.memWrite(0x10,(byte)0b11000000);
+        nes.loadAndRun(new int[]{
+           0x24,
+           0x10,
+           0x00
         });
         assertTrue((nes.status & 0b10000000) != 0); // Negative flag should be set
         assertTrue((nes.status & 0b01000000) != 0); // Overflow flag should be set
@@ -117,11 +117,11 @@ public class DemoNESTest {
         DemoNES cpu = new DemoNES();
         cpu.status |= DemoNES.NEGATIVE;  // set negative
 
-        cpu.loadAndRun(new byte[]{
-            (byte)0x30,  // BMI
-            (byte)0x01,  // +1
-            (byte)0x00,  // skipped
-            (byte)0x00   // executed
+        cpu.loadAndRun(new int[]{
+           0x30,  // BMI
+           0x01,  // +1
+           0x00,  // skipped
+           0x00   // executed
         });
         
         assertEquals(0x8003, cpu.programCounter);
@@ -131,11 +131,11 @@ public class DemoNESTest {
         DemoNES cpu = new DemoNES();
         cpu.status &= ~DemoNES.ZERO;  // clear zero (so branch happens)
 
-        cpu.loadAndRun(new byte[]{
-            (byte)0xD0,  // BNE
-            (byte)0x01,  // +1
-            (byte)0x00,  // skipped
-            (byte)0x00   // executed
+        cpu.loadAndRun(new int[]{
+           0xD0,  // BNE
+           0x01,  // +1
+           0x00,  // skipped
+           0x00   // executed
         });
 
         assertEquals(0x8004, cpu.programCounter);
@@ -143,7 +143,7 @@ public class DemoNESTest {
     @Test
     void test_bne_not_taken() {
         DemoNES cpu = new DemoNES();
-        cpu.load(new byte[]{(byte)0xD0, (byte)0x01, (byte)0x00});
+        cpu.load(new int[]{(byte)0xD0,0x01,0x00});
         cpu.reset();
         cpu.status |= DemoNES.ZERO;  // set AFTER reset
         cpu.run();
@@ -152,11 +152,11 @@ public class DemoNESTest {
     @Test
     void test_bpl_branch_taken() {
         DemoNES cpu = new DemoNES();
-        cpu.load(new byte[]{
-            (byte)0x10,  // BPL
-            (byte)0x01,  // +1
-            (byte)0x00,  // skipped
-            (byte)0x00   // executed
+        cpu.load(new int[]{
+           0x10,  // BPL
+           0x01,  // +1
+           0x00,  // skipped
+           0x00   // executed
         });
         cpu.reset();
         cpu.status &= ~DemoNES.NEGATIVE;  // clear negative (so branch happens)
@@ -166,11 +166,11 @@ public class DemoNESTest {
     @Test
     void test_bpl_no_branch_taken() {
         DemoNES cpu = new DemoNES();
-        cpu.load(new byte[]{
-            (byte)0x10,  // BPL
-            (byte)0x01,  // +1
-            (byte)0x00,  // skipped
-            (byte)0x00   // executed
+        cpu.load(new int[]{
+           0x10,  // BPL
+           0x01,  // +1
+           0x00,  // skipped
+           0x00   // executed
         });
         cpu.reset();
         cpu.status |= DemoNES.NEGATIVE;  // set negative (so branch not taken)
@@ -181,9 +181,9 @@ public class DemoNESTest {
     void test_cmp_equal() {
         DemoNES cpu = new DemoNES();
 
-        cpu.load(new byte[]{
-            (byte)0xC9, 0x42,
-            (byte)0x00
+        cpu.load(new int[]{
+           0xC9, 0x42,
+           0x00
         });
         cpu.reset();
         cpu.registerA = 0x42;  // set AFTER reset
@@ -197,11 +197,11 @@ public class DemoNESTest {
     void test_jmp_absolute() {
         DemoNES cpu = new DemoNES();
 
-        cpu.loadAndRun(new byte[]{
-            (byte)0x4C, 0x05, (byte)0x80,  // JMP $8005
-            (byte)0x00,                   // skipped
-            (byte)0x00,                   // skipped
-            (byte)0x00                    // executed
+        cpu.loadAndRun(new int[]{
+           0x4C, 0x05,0x80,  // JMP $8005
+           0x00,                   // skipped
+           0x00,                   // skipped
+           0x00                    // executed
         });
 
         assertEquals(0x8006, cpu.programCounter);
@@ -213,11 +213,11 @@ public class DemoNESTest {
 
         cpu.memWriteU16(0x9000, 0x8005);
 
-        cpu.loadAndRun(new byte[]{
-            (byte)0x6C, 0x00, (byte)0x90,  // JMP ($9000)
-            (byte)0x00,                   // skipped
-            (byte)0x00,                   // skipped
-            (byte)0x00                    // executed
+        cpu.loadAndRun(new int[]{
+           0x6C, 0x00,0x90,  // JMP ($9000)
+           0x00,                   // skipped
+           0x00,                   // skipped
+           0x00                    // executed
         });
 
         assertEquals(0x8006, cpu.programCounter);
@@ -227,11 +227,11 @@ public class DemoNESTest {
     void test_jsr_rts() {
         DemoNES cpu = new DemoNES();
 
-        cpu.loadAndRun(new byte[]{
-            (byte)0x20, 0x05, (byte)0x80,  // JSR $8005
-            (byte)0x00,                   // BRK (should run after RTS)
-            (byte)0x00,                   // padding
-            (byte)0x60                    // RTS at $8005
+        cpu.loadAndRun(new int[]{
+           0x20, 0x05,0x80,  // JSR $8005
+           0x00,                   // BRK (should run after RTS)
+           0x00,                   // padding
+           0x60                    // RTS at $8005
         });
 
         assertEquals(0x8004, cpu.programCounter);
@@ -241,9 +241,9 @@ public class DemoNESTest {
     void test_pha() {
         DemoNES cpu = new DemoNES();
 
-        cpu.load(new byte[]{
-            (byte)0x48,
-            (byte)0x00
+        cpu.load(new int[]{
+           0x48,
+           0x00
         });
         cpu.reset();
         cpu.registerA = 0x42;  // set AFTER reset
@@ -257,9 +257,9 @@ public class DemoNESTest {
     void test_php() {
         DemoNES cpu = new DemoNES();
 
-        cpu.load(new byte[]{
-            (byte)0x08,  // PHP
-            (byte)0x00   // BRK
+        cpu.load(new int[]{
+           0x08,  // PHP
+           0x00   // BRK
         });
         cpu.reset();
         cpu.status = 0x42;  // set AFTER reset
